requirements.txt
aniso8601==8.1.0
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-Testing==0.8.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
packaging==20.8
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
pandas==1.2.0
pytest_asyncio==0.14.0
APScheduler==3.7.0
jwt==1.1.0